feat: Initial implementation of iframe protocol
The iframe protocol allows crowdfiller to be embedded in another
web applications.

It's a simple protocol that only has 3 messages:

* CROWDFILLER_DO_SUBMIT: fired by the main page when a submit is
  required.
* CROWDFILLER_SUBMITTED: fired by the iframe if the response was
  processed by the server successfully.
* CROWDFILLER_VALIDATION_ERROR: fired by the iframe if the response
  contains validation errors.

This commit also add support for the 'iframe' flag on the response url
that enables the iframe protocol, in a future we can automatically
detect if we are in a iframe an this flag will be ignored.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>